Be/refactor/modify record

See merge request s09-final/S09P31C103!46

+++ back-end/src/main/java/com/ssafy/eoullim/model/Room.java
@@ -10,8 +10,8 @@ import java.util.List;
 public class Room {
     private String sessionId;
     private String recordingId;
-    private Integer childOne;
-    private Integer childTwo;
+    private Long childOne;
+    private Long childTwo;
     private List<Integer> random;
     private String guideSeq;
     private String timeline;
